+++ reporters/summary_reporter.py
@@ -1,5 +1,0 @@
-"""Compatibility wrapper for ``analyzer.reporters.summary_reporter``."""
-
-from analyzer.reporters.summary_reporter import SummaryReporter
-
-__all__ = ["SummaryReporter"]